Refactor components for cleaner JSX structure and improved readability

+++ src/components/DeleteBtn.jsx
@@ -1,17 +1,13 @@
 const DeleteBtn = ({ iceCream, onDeleteOne }) => {
   const handleClick = () => {
-    onDeleteOne(iceCream);
+    const response = confirm(
+      `Deleting ${iceCream.brand} ${iceCream.flavour} ice cream from the list. Continue?`,
+    );
+    if (response) {
+      onDeleteOne(iceCream);
+    }
   };
 
-  return (
-    <>
-      <button
-        className="del"
-        onClick={handleClick}
-      >
-        &#128465;
-      </button>
-    </>
-  );
+  return <button onClick={handleClick}>&#128465;</button>;
 };
 export default DeleteBtn;
